refactor: rewrite test for empty array
Every test has a cost to maintain. As a result, the goal is not to write
as many tests as possible but rather write just enough tests.

Inductive test cases are one of way of structuring your tests. E.g. in
this case we write a test case for the base case, the empty array and
one test case for the inductive step which is a non-empty array.